chore: move `page-header` and `toaster` to core
